@@ -4,16 +4,20 @@
   "description": "an example of react.js in server",
   "main": "server.js",
   "scripts": {
+    "build": "babel -d . src/",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "ruanyf",
   "license": "ISC",
   "dependencies": {
-    "browserify": "~9.0.3",
+    "browserify": "^16.2.3",
     "literalify": "~0.4.0",
-    "react": "~0.13.1"
+    "react": "^16.7.0",
+    "react-dom": "^16.7.0"
   },
   "devDependencies": {
-    "react-tools": "~0.13.1"
+    "babel-cli": "^6.26.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1"
   }
 }
