@@ -1,9 +1,11 @@
 var http = require('http'),
     browserify = require('browserify'),
     literalify = require('literalify'),
-    React = require('react');
+    React = require('react'),
+    ReactDOMServer = require('react-dom/server');
 
-var App = require('./app');
+// export default
+import App from './app';
 
 http.createServer(function(req, res) {
   if (req.url == '/') {
@@ -14,16 +16,17 @@ http.createServer(function(req, res) {
         'Item 1'
       ]
     };
-    var html = React.renderToStaticMarkup(
+    var html = ReactDOMServer.renderToStaticMarkup(
       <body>
         <div id="content" dangerouslySetInnerHTML={{__html:
-          React.renderToString(<App items={props.items}/>)
-        }} />,
+          ReactDOMServer.renderToString(<App items={props.items}/>)
+        }} />
 
         <script dangerouslySetInnerHTML={{__html:
         'var APP_PROPS = ' + JSON.stringify(props) + ';'
         }}/>
-        <script src="//fb.me/react-0.13.1.min.js"/>
+        <script src="https://cdn.jsdelivr.net/npm/react@16.7.0/umd/react.production.min.js"/>
+        <script src="https://cdn.jsdelivr.net/npm/react-dom@16.7.0/umd/react-dom.production.min.js"/>
         <script src="/bundle.js"/>
       </body>
     );
@@ -33,10 +36,12 @@ http.createServer(function(req, res) {
     res.setHeader('Content-Type', 'text/javascript');
     browserify()
       .add('./browser.js')
-      .transform(literalify.configure({react: 'window.React'}))
+      .transform(literalify.configure({
+        'react': 'window.React',
+        'react-dom': 'window.ReactDOM',
+      }))
       .bundle()
       .pipe(res);
-
   } else {
     res.statusCode = 404;
     res.end();
